# agent/openapi.yaml
openapi: 3.1.0
info:
  title: Daftar Closing Agent — Device ↔ Cloud Run Wire Contract
  version: 2.4.1
  summary: Frozen Appendix J contract for All Things Agentic (Submission Period 2026).
  description: |
    Normative machine-readable contract for Stages 1–4 of `docs/roadmap_v2.md` **Appendix J**.
    Stage 4.7 adds deterministic `POST /v1/email/send-batch` (J.7) beside ADK `/run`.
    Stage 5 speech quality adds unary `POST /v1/tts` (Chirp 3 HD Enceladus MP3; not Gemini Live).

    ## Architecture
    - Cloud Run hosts a Google **ADK** agent (Python) that emits **proposals**, never Drift writes.
    - Flutter commits money/create/WhatsApp prepare only after an on-device **confirm gate**.
    - Model: `gemini-3.5-flash` on Vertex (`GOOGLE_GENAI_USE_VERTEXAI=TRUE`,
      `GOOGLE_CLOUD_LOCATION=global` or `us`). Gemini 3.1 product models are banned for contest.

    ## Context carriage (locked)
    Appendix J request fields travel as JSON under ADK session state key **`daftarContext`**
    (`DeviceAgentRequest`) on session create and/or `POST /run` `stateDelta`.
    The user utterance is ADK `newMessage.parts[].text` and MUST equal `goalText`.

    ## Money (HARD)
    All monetary amounts are **`amountMinor: integer`** (smallest currency unit) plus
    `currencyCode: string`. Never `number` / float / double / Decimal / string amounts.

    ## localDay (J.4)
    1. Persist txn `createdAt` in UTC (existing Drift invariant).
...

    ## Auth (J.1)
...
      treat it as live OR-auth.
...

...

...
servers:
...
security:
  - GoogleIdToken: []
paths:
...
  /v1/email/send-batch:
    post:
      tags: [email-send]
      operationId: sendEmailBatch
      summary: Deterministic Gmail SMTP send-batch (Appendix J.7)
      description: |
        Multipart manifest + optional per-contact PDF (`pdf_{contactId}`). Server
        assembles Appendix C.2 `text/plain` from named params (never LLM `body`,
        never `amountMinor`). Missing PDF part → text-only SMTP. One SMTP
        transaction per recipient. Cap 20. Auth is Cloud Run IAM + J.1 Google ID
        token (same custom audiences as `/run`). Do not add a second in-process
        JWT check. Shared-secret fallback is unimplemented.
      security:
        - GoogleIdToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest:
                  type: string
                  description: JSON EmailSendBatchManifest (see components.schemas)
              additionalProperties:
                type: string
                format: binary
                description: Optional PDF part named `pdf_{contactId}` (`application/pdf`, ≤5 MB). Missing → text-only SMTP.
            encoding:
              manifest:
                contentType: application/json
      responses:
        "200":
          description: Per-recipient SMTP results (recipient order)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailSendBatchResponse"
        "400":
          description: Invalid manifest (cap >20, UUID, locale, missing required strings)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    GoogleIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Google OAuth **ID token** for the signed-in merchant. Audience is Cloud Run
        **custom audiences** (Web + Android + iOS OAuth client IDs). GSI `aud` is
        the Web client ID; PKCE `aud` is the installed-app client ID — not the
        `*.run.app` service URL. `gcloud auth print-identity-token` uses the
        service URL (also accepted). Primary auth for Flutter → Cloud Run
        (Appendix J.1).
    AgentSharedSecret:
      type: apiKey
      in: header
      name: X-Daftar-Agent-Secret
      description: |
        **Not implemented.** Documented historically as a demo fallback. Do not
        send this header. Live auth is Google ID token + Cloud Run IAM only.
...
  responses:
    Unauthorized:
      description: Missing or invalid identity (Cloud Run IAM / token)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
    Uuid:
      type: string
      format: uuid
      pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
...
    EmailSendRowResult:
      type: object
      additionalProperties: false
      required: [contactId, toMasked, status]
      properties:
        contactId:
          $ref: "#/components/schemas/Uuid"
        toMasked:
          type: string
          description: Masked recipient (local-part last-4 or m***@domain)
        status:
          type: string
          enum: [sent, failed, skippedDuplicate]
        smtpMessageId:
          oneOf:
            - type: string
            - type: "null"
          description: Present iff status is sent
        smtpCode:
          oneOf:
            - type: integer
            - type: "null"
        smtpMessage:
          oneOf:
            - type: string
            - type: "null"
          description: Safe summary; never the App Password
    EmailSendBatchResponse:
      type: object
      additionalProperties: false
      required: [batchId, results, needsHuman]
      properties:
        batchId:
          $ref: "#/components/schemas/Uuid"
        results:
          type: array
          items:
            $ref: "#/components/schemas/EmailSendRowResult"
        needsHuman:
          type: boolean
          description: true on SMTP 535 / From mismatch / secret unavailable
...